restructure and add basic docs
